reorganize tests
