--- test/parallel/test-path-glob.js
@@ -1,77 +1,0 @@
-'use strict';
-
-require('../common');
-const assert = require('assert');
-const path = require('path');
-
-function testMatch() {
-  const globs = {
-    win32: [
-      ['foo\\bar\\baz', 'foo\\[bcr]ar\\baz', true], // Matches 'bar' or 'car' in 'foo\\bar'
-      ['foo\\bar\\baz', 'foo\\[!bcr]ar\\baz', false], // Matches anything except 'bar' or 'car' in 'foo\\bar'
-      ['foo\\bar\\baz', 'foo\\[bc-r]ar\\baz', true], // Matches 'bar' or 'car' using range in 'foo\\bar'
-      ['foo\\bar\\baz', 'foo\\*\\!bar\\*\\baz', false], // Matches anything with 'foo' and 'baz' but not 'bar'
-      // in between
-      ['foo\\bar1\\baz', 'foo\\bar[0-9]\\baz', true], // Matches 'bar' followed by any digit in 'foo\\bar1'
-      ['foo\\bar5\\baz', 'foo\\bar[0-9]\\baz', true], // Matches 'bar' followed by any digit in 'foo\\bar5'
-      ['foo\\barx\\baz', 'foo\\bar[a-z]\\baz', true], // Matches 'bar' followed by any lowercase letter in 'foo\\barx'
-      ['foo\\bar\\baz\\boo', 'foo\\[bc-r]ar\\baz\\*', true], // Matches 'bar' or 'car' in 'foo\\bar'
-      ['foo\\bar\\baz', 'foo/**', true], // Matches anything in 'foo'
-      ['foo\\bar\\baz', '*', false], // No match
-    ],
-    posix: [
-      ['foo/bar/baz', 'foo/[bcr]ar/baz', true], // Matches 'bar' or 'car' in 'foo/bar'
-      ['foo/bar/baz', 'foo/[!bcr]ar/baz', false], // Matches anything except 'bar' or 'car' in 'foo/bar'
-      ['foo/bar/baz', 'foo/[bc-r]ar/baz', true], // Matches 'bar' or 'car' using range in 'foo/bar'
-      ['foo/bar/baz', 'foo/*/!bar/*/baz', false], // Matches anything with 'foo' and 'baz' but not 'bar' in between
-      ['foo/bar1/baz', 'foo/bar[0-9]/baz', true], // Matches 'bar' followed by any digit in 'foo/bar1'
-      ['foo/bar5/baz', 'foo/bar[0-9]/baz', true], // Matches 'bar' followed by any digit in 'foo/bar5'
-      ['foo/barx/baz', 'foo/bar[a-z]/baz', true], // Matches 'bar' followed by any lowercase letter in 'foo/barx'
-      ['foo/bar/baz/boo', 'foo/[bc-r]ar/baz/*', true], // Matches 'bar' or 'car' in 'foo/bar'
-      ['foo/bar/baz', 'foo/**', true], // Matches anything in 'foo'
-      ['foo/bar/baz', '*', false], // No match
-    ],
-  };
-
-
-  for (const [platform, platformGlobs] of Object.entries(globs)) {
-    for (const [pathStr, glob, expected] of platformGlobs) {
-      const actual = path[platform].matchesGlob(pathStr, glob);
-      assert.strictEqual(actual, expected, `Expected ${pathStr} to ` + (expected ? '' : 'not ') + `match ${glob} on ${platform}`);
-    }
-  }
-
-  // Test for non-string input
-  assert.throws(() => path.matchesGlob(123, 'foo/bar/baz'), /.*must be of type string.*/);
-  assert.throws(() => path.matchesGlob('foo/bar/baz', 123), /.*must be of type string.*/);
-}
-
-function testEscaping() {
-  const samples = [
-    ['file*.txt', 'file[*].txt'],
-    ['file?.txt', 'file[?].txt'],
-    ['file[abc].txt', 'file[[]abc[]].txt'],
-    ['folder/**/*.txt', 'folder/[*][*]/[*].txt'],
-    ['file{1,2}.txt', 'file[{]1,2[}].txt'],
-    ['file[0-9]?.txt', 'file[[]0-9[]][?].txt'],
-    ['C:\\Users\\*.txt', 'C:\\Users\\[*].txt'],
-    ['?[]', '[?][[][]]'],
-  ];
-
-  for (const [pattern, expected] of samples) {
-    const actual = path.escapeGlob(pattern);
-    assert.strictEqual(actual, expected, `Expected ${pattern} to be escaped as ${expected}, got ${actual} instead`);
-  }
-
-  for (const [expected, escapedPattern] of samples) {
-    const actual = path.unescapeGlob(escapedPattern);
-    assert.strictEqual(actual, expected, `Expected ${escapedPattern} to be unescaped as ${expected}, got ${actual} instead`);
-  }
-
-  // Test for non-string input
-  assert.throws(() => path.escapeGlob(123), /.*must be of type string.*/);
-  assert.throws(() => path.unescapeGlob(123), /.*must be of type string.*/);
-}
-
-testMatch();
-testEscaping();
